refactor: add error handling

+++ src/commands/files/mv-command.js
@@ -1,18 +1,19 @@
 import path from 'path';
-import { copyFile, unlink } from 'node:fs/promises';
+import { createReadStream, createWriteStream } from 'fs';
+import { stat, unlink } from 'fs/promises';
 
 import { getWorkingDirectory, stdoutText, write } from '../../utils/index.js';
 import { AbstractCommand } from '../abstract-command.js';
 
 export class MvCommand extends AbstractCommand {
   #filePath = '';
-  #newFilePath = '';
+  #newDir = '';
 
-  constructor([filePath, newFilePath]) {
+  constructor([filePath, newDir]) {
     super();
 
     this.#filePath = filePath;
-    this.#newFilePath = newFilePath;
+    this.#newDir = newDir;
   }
 
   static get commandName() {
@@ -20,12 +21,26 @@ export class MvCommand extends AbstractCommand {
   }
 
   async execute() {
+    const workingDirectory = getWorkingDirectory();
+    const filePath = path.resolve(workingDirectory, this.#filePath);
+    const fileName = path.basename(filePath);
+    const newDir = path.resolve(workingDirectory, this.#newDir, fileName);
+
     try {
-      const workingDirectory = getWorkingDirectory();
-      const filePath = path.resolve(workingDirectory, this.#filePath);
-      const newFilePath = path.resolve(workingDirectory, this.#newFilePath);
-      await copyFile(filePath, newFilePath);
-      await unlink(filePath);
+      const stats = await stat(filePath);
+      if (stats.isDirectory()) throw new Error('cp: cannot copy directories');
+      const readableStream = createReadStream(filePath);
+      const writeStream = createWriteStream(newDir);
+      return await new Promise((resolve, reject) => {
+        readableStream.on('error', reject);
+        readableStream.on('end', async () => {
+          await unlink(filePath);
+          resolve();
+        });
+        writeStream.on('error', reject);
+        writeStream.on('finish', resolve);
+        readableStream.pipe(writeStream);
+      });
     } catch (err) {
       write(stdoutText.sayFailed());
     }
